bug fixed

+++ src/java/DataModel/LocationData/FavLocation.java
@@ -3,8 +3,8 @@ package DataModel.LocationData;
 public class FavLocation {
     String type;
     Location location;
-    float latitude;
-    float longitude;
+    double latitude;
+    double longitude;
 
     public FavLocation(String type, Location location) {
         this.type = type;
@@ -16,19 +16,19 @@ public class FavLocation {
     public boolean containLocation(Location location){
         return location.getName().compareToIgnoreCase(this.location.getName()) == 0 && location.getLatitude() == this.latitude && location.getLongitude() == this.longitude ;
     }
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
