+++ src/java/DataModel/LocationData/Restaurant.java
@@ -1,11 +1,9 @@
 package DataModel.LocationData;
 
-import DataModel.LocationData.Location;
-
 public class Restaurant extends Location {
     String[] characteristics;
 
-    public Restaurant(String name, float[] coordinates, String description, String[] characteristics) {
+    public Restaurant(String name, double[] coordinates, String description, String[] characteristics) {
         super(name, coordinates, description);
         this.characteristics = characteristics;
     }
